requirements.txt
# ============================
# LightenDiffusion Dependencies
# ============================

# Core deep learning
torch>=1.12.0
torchvision>=0.13.0

# Image processing
opencv-python>=4.5.0
Pillow>=9.0.0
kornia>=0.6.7
einops>=0.6.0

# Utilities
numpy>=1.21.0
scipy>=1.7.0
PyYAML>=6.0
tqdm>=4.64.0

# Logging & visualization
tensorboard>=2.10.0

# Metrics
torchmetrics>=0.11.0
lpips>=0.1.4

# Optimization
timm>=0.6.0

# Training acceleration
accelerate>=0.15.0